+++ paper/src/main/java/com/badbones69/crazycrates/CrazyCrates.java
@@ -8,10 +8,10 @@ import com.badbones69.crazycrates.listeners.menus.CrateAdminListener;
 import com.badbones69.crazycrates.listeners.menus.CrateMenuListener;
 import com.badbones69.crazycrates.listeners.menus.CratePreviewListener;
 import com.badbones69.crazycrates.listeners.platforms.PaperListener;
-import com.badbones69.crazycrates.managers.BukkitUserManager;
-import com.badbones69.crazycrates.managers.crates.CrateManager;
-import com.badbones69.crazycrates.managers.crates.other.quadcrates.SessionManager;
-import com.badbones69.crazycrates.utils.MsgUtils;
+import com.badbones69.crazycrates.tasks.BukkitUserManager;
+import com.badbones69.crazycrates.tasks.crates.CrateManager;
+import com.badbones69.crazycrates.tasks.crates.other.quadcrates.SessionManager;
+import com.badbones69.crazycrates.api.utils.MsgUtils;
 import org.bukkit.plugin.java.JavaPlugin;
 import com.badbones69.crazycrates.common.config.types.ConfigKeys;
 import com.badbones69.crazycrates.api.FileManager;
