+++ paper/src/main/java/com/badbones69/crazycrates/api/objects/Crate.java
@@ -5,8 +5,8 @@ import org.bukkit.SoundCategory;
 import us.crazycrew.crazycrates.api.enums.types.CrateType;
 import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.FileManager;
-import com.badbones69.crazycrates.managers.crates.other.CosmicCrateManager;
-import com.badbones69.crazycrates.managers.crates.other.AbstractCrateManager;
+import com.badbones69.crazycrates.tasks.crates.other.CosmicCrateManager;
+import com.badbones69.crazycrates.tasks.crates.other.AbstractCrateManager;
 import org.jetbrains.annotations.NotNull;
 import com.badbones69.crazycrates.common.crates.CrateHologram;
 import de.tr7zw.changeme.nbtapi.NBTItem;
@@ -17,10 +17,10 @@ import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
-import com.badbones69.crazycrates.managers.InventoryManager;
+import com.badbones69.crazycrates.tasks.InventoryManager;
 import com.badbones69.crazycrates.api.builders.types.CratePreviewMenu;
-import com.badbones69.crazycrates.utils.MiscUtils;
-import com.badbones69.crazycrates.utils.MsgUtils;
+import com.badbones69.crazycrates.api.utils.MiscUtils;
+import com.badbones69.crazycrates.api.utils.MsgUtils;
 
 import java.io.File;
 import java.io.IOException;
