+++ paper/src/main/java/com/badbones69/crazycrates/listeners/crates/QuadCrateListener.java
@@ -9,15 +9,14 @@ import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.inventory.meta.ItemMeta;
 import com.badbones69.crazycrates.CrazyCrates;
-import com.badbones69.crazycrates.managers.crates.other.quadcrates.QuadCrateManager;
-import com.badbones69.crazycrates.managers.crates.other.quadcrates.SessionManager;
+import com.badbones69.crazycrates.tasks.crates.other.quadcrates.QuadCrateManager;
+import com.badbones69.crazycrates.tasks.crates.other.quadcrates.SessionManager;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.ItemBuilder;
 import com.badbones69.crazycrates.api.objects.Prize;
 import com.badbones69.crazycrates.api.enums.PersistentKeys;
 import com.badbones69.crazycrates.api.enums.Messages;
 import com.badbones69.crazycrates.support.structures.blocks.ChestManager;
-import org.bukkit.Sound;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Item;
