--- paper/src/main/java/com/badbones69/crazycrates/managers/crates/other/AbstractCrateManager.java
@@ -1,3 +1,0 @@
-package com.badbones69.crazycrates.managers.crates.other;
-
-public abstract class AbstractCrateManager {}
