+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/other/AbstractCrateManager.java
@@ -1,0 +1,3 @@
+package com.badbones69.crazycrates.tasks.crates.other;
+
+public abstract class AbstractCrateManager {}
