+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/types/QuadCrate.java
@@ -1,4 +1,4 @@
-package com.badbones69.crazycrates.managers.crates.types;
+package com.badbones69.crazycrates.tasks.crates.types;
 
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.CrateLocation;
@@ -10,8 +10,8 @@ import com.badbones69.crazycrates.common.crates.quadcrates.CrateSchematic;
 import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.builders.CrateBuilder;
 import com.badbones69.crazycrates.support.structures.StructureHandler;
-import com.badbones69.crazycrates.managers.crates.CrateManager;
-import com.badbones69.crazycrates.managers.crates.other.quadcrates.QuadCrateManager;
+import com.badbones69.crazycrates.tasks.crates.CrateManager;
+import com.badbones69.crazycrates.tasks.crates.other.quadcrates.QuadCrateManager;
 import java.util.Random;
 
 public class QuadCrate extends CrateBuilder {
